I've moved around the original files so we can start having a bit of organization, though I just put them in those folders so that they would be separated from the things we continue adding, until we can figure out the correct place to put them.
I've also added some of the map generation stuff, which is incomplete, but
I'm putting it here for our coming meeting.  It should compile and/or not
show any errors on Eclipse if you have slick2d set up correctly.

Lastly, I've renamed the Map class to DungeonMap to be more specific.