Small Update to ignore dungeon maps and clean up
I've added a little bit more to the MapUtil class so that dungeon maps
are created with a .dm file extension.  Then I made it so that the
.gitignore file will ignore it.  I deleted the MapWriter file because it
is old and useless now.

I am deleting the "test" package as well, since I only created it so
that I would still have a working original at the same time as when I
was messing with that code.  Probably not needed anymore, since you've
merged things.  We can always look back later if needed.