Part of Doors and more
I added some code into DungeonMapGenerator to create doors, and
reorganized a bit on how a DungeonMap is created.  I also created a Door
class and a Teleporter class (Teleporter was what I was thinking about
originally, but it turned out a mess that way, so a Teleporter is simply
a teleport to somewhere else on the map)

Added in another image set just because it has a door, we can change it
later.

I also updated the image parse file so that we can read tiles from
multiple files.

Other misc. things as well.