Added a Name Tag to the Player, Smarter Monster
Before, the monster could take a step and at the end of the step, the
player would be at the edge of its vision.  However, because the player
moves first, the player would step out of vision, and the monster would
not even know that it had "seen" the player.  To fix this, at the end of
the monster's move, it tries to detect again and come up with another
plan.  So even if the player moves away, the monster will have a plan to
move closer.  And after it moves, it will see the player again at the
edge of its vision.

Also added a name tag to the player, just to test out how.