Added in test ability to see monster's LOS
Also tried to include detection, but it is bugged right now, or else the
monster sucks at detecting me.  We probably need to include some sort of
awareness feature, because it doesn't make sense for it to lose track of
me once it has discovered me in the first place.

Add whatever you want to the TestUtil class, so that you can have on/off
switches to test features, or dev tools, etc.

Also added a few methods into OSC that simply call the methods in its
base creature.